+++ packages/kit/test/build-errors/apps/prerenderable-incorrect-fragment/svelte.config.js
@@ -1,9 +1,9 @@
+import adapter from '../../../../../adapter-auto/index.js';
+
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
-		prerender: {
-			handleHttpError: 'warn'
-		}
+		adapter: adapter()
 	}
 };